requirements.txt
crewai>=1.11.0
langchain-anthropic>=1.4.0
fastapi>=0.135.0
uvicorn[standard]>=0.42.0
pydantic-settings>=2.10.0
apscheduler>=3.10.0
docker>=7.0.0
chromadb>=0.4.0
sentence-transformers>=5.0.0
trafilatura>=2.0.0
youtube-transcript-api>=1.2.0
python-dotenv>=1.0.0
beautifulsoup4>=4.12.0
requests>=2.31.0
firebase-admin>=6.5.0
pypdf>=4.0,<6
python-docx>=1.0
openpyxl>=3.1
Pillow>=10.0
litellm>=1.50.0
mem0ai>=0.1.0
